Try fixing GitHub Actions
